Extract collectors to collector.js
