+++ codemod/src/utils.js
@@ -2,8 +2,6 @@ const recast = require('recast');
 
 const Files = require('./files');
 
-const { IgnoredNewExpressions } = require('./constants');
-
 const getFileFunctions = file => Files
   .getFileAST(file)
   .map(func => func.id.name);
@@ -12,63 +10,6 @@ const getConstants = () => Files
   .getFileAST('Constants')
   .map(variable => variable.declarations[0])
   .map(variable => variable.id.name);
-
-function collectIdentifiers(ast, identifiers) {
-  const existingIdentifiers = new Set();
-
-  function visitIdentifier(path) {
-    const { name } = path.value;
-
-    if (!identifiers) {
-      existingIdentifiers.add(name)
-    } else {
-      const isIdentifier = identifiers.includes(name);
-      isIdentifier && existingIdentifiers.add(name)
-    }
-    
-    return this.traverse(path);
-  }
-
-  recast.visit(ast, { visitIdentifier })
-
-  return Array.from(existingIdentifiers);
-}
-
-function collectClassesDeclaration(ast) {
-  const declaredClasses = new Set();
-
-  function visitClassDeclaration(path) {
-    declaredClasses.add(path.value)
-
-    return this.traverse(path);
-  }
-
-  recast.visit(ast, { visitClassDeclaration }) 
-
-  return Array.from(declaredClasses).filter(Boolean);
-}
-
-function collectClassesUsage(ast) {
-  const usedClasses = new Set();
-  
-  function visitClassDeclaration(path) {
-    if (!IgnoredNewExpressions.includes(path.value.id.name)) {
-      const { superClass } = path.value;
-      superClass && usedClasses.add(superClass.name);
-    }
-
-    return this.traverse(path);
-  }
-
-  function visitNewExpression(path) {
-    usedClasses.add(path.value.callee.name)
-    return this.traverse(path);
-  }
-
-  recast.visit(ast, { visitClassDeclaration }) 
-
-  return Array.from(usedClasses).filter(Boolean);
-}
 
 const buildMemberExpression = config => new recast.types.NodePath(
   recast.types.builders.memberExpression(
@@ -130,8 +71,5 @@ function addGlobalPrefix(fileAST) {
 
 module.exports.addGlobalPrefix = addGlobalPrefix;
 
-module.exports.collectIdentifiers = collectIdentifiers;
-module.exports.collectClassesDeclaration = collectClassesDeclaration;
-module.exports.collectClassesUsage = collectClassesUsage;
 module.exports.getFileFunctions = getFileFunctions;
 module.exports.getConstants = getConstants;
